requirements.txt
pandas
mlflow==3.0.1
notebook
numpy
scikit-learn
matplotlib
python-box==6.0.2
pyYAML
tqdm
ensure==1.0.4
joblib
types-pyYAML
Flask
Flask-Cors
-e .

